backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
openai>=1.55.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
